+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- N/A
+- Wait for Zwift to start on Windows and Linux
 
 ## [0.1.0] - 2025-08-02